requirements.txt
python-dotenv>=1.0.0
flake8>=6.0.0
black>=23.12.1
pytest>=8.0.0
mlflow==2.14.0          # MLflow固定版本
scikit-learn==1.3.2     # 机器学习库（鸢尾花模型用）
pandas==2.1.4
numpy==1.26.3
setuptools<81.0 